// cloudreve-official-frontend/src/api/explorer.ts
export const FileType = {
  file: 0,
  folder: 1,
};

export interface FileResponse {
  type: number;
  id: string;
  name: string;
  permission?: string;
  created_at: string;
  updated_at: string;
  size: number;
  metadata?: {
    [key: string]: string;
  };
  path: string;
  shared?: boolean;
  capability?: string;
  owned?: boolean;
  folder_summary?: FolderSummary;
  extended_info?: ExtendedInfo;
  primary_entity?: string;
}

export interface FolderSummary {
  size: number;
  files: number;
  folders: number;
  completed: boolean;
  calculated_at: string;
}

export interface ExtendedInfo {
  storage_policy?: StoragePolicy;
  storage_policy_inherited: boolean;
  storage_used: number;
  shares?: Share[];
  entities?: Entity[];
  permissions?: PermissionSettingReq;
  view?: ExplorerView;
  direct_links?: DirectLink[];
}

export interface DirectLink {
  id: string;
  created_at: string;
  url: string;
  downloaded: number;
}

export interface Entity {
  id: string;
  type: number;
  created_at: string;
  storage_policy?: StoragePolicy;
  size: number;
  created_by?: User;
  encrypted_with?: EncryptionCipher;
}

export interface Share {
  id: string;
  name?: string;
  expires?: string;
  is_private?: boolean;
  share_view?: boolean;
  remain_downloads?: number;
  created_at?: string;
  price?: number;
  permission_setting?: PermissionSettingReq;
  url: string;
  visited: number;
  downloaded: number;
  expired?: boolean;
  unlocked: boolean;
  password_protected: boolean;
  source_type?: number;
  owner: User;
  permissions?: string;
  source_uri?: string;
  password?: string;
  show_readme?: boolean;
}

export enum PolicyType {
  local = "local",
  remote = "remote",
  oss = "oss",
  qiniu = "qiniu",
  onedrive = "onedrive",
  cos = "cos",
  upyun = "upyun",
  s3 = "s3",
  ks3 = "ks3",
  obs = "obs",
  load_balance = "load_balance",
}

export interface StoragePolicy {
  id: string;
  name: string;
  allowed_suffix?: string[];
  denied_suffix?: string[];
  allowed_name_regexp?: string;
  denied_name_regexp?: string;
  max_size: number;
  type: PolicyType;
  relay?: boolean;
  weight?: number;
  children?: StoragePolicy[];
  chunk_concurrency?: number;
  encryption?: boolean;
  streaming_encryption?: boolean;
}

export interface PaginationResults {
  page: number;
  page_size: number;
  total_items?: number;
  next_token?: string;
  is_cursor?: boolean;
}

export interface NavigatorProps {
  capability?: string;
  max_page_size: number;
  order_by_options: string[];
  order_direction_options: string[];
}

export interface ExplorerView {
  page_size: number;
  order?: string;
  order_direction?: string;
  view?: string;
  thumbnail?: boolean;
  columns?: ListViewColumnSetting[];
  gallery_width?: number;
}
// ...
export interface PermissionSettingReq {
  same_group?: string;
  everyone?: string;
  other?: string;
  anonymous?: string;
  group_explicit?: {
    [key: string]: string;
  };
  user_explicit?: {
    [key: string]: string;
  };
}
// ...
export interface LogEntry {
  category: number;
  failed?: boolean;
  error?: string;
  user_agent?: string;
  is_system?: boolean;
  email_to?: string;
  email_title?: string;
  original_name?: string;
  new_name?: string;
  from?: string;
  to?: string;
  entity_create_time?: string;
  storage_policy?: string;
  storage_policy_id?: string;
  direct_link_id?: string;
}

export interface FileActivity {
  content: LogEntry;
  created_at: string;
  user?: User;
  id: string;
  version_id?: string;
}

export interface FileActivitiesResponse {
  activities: FileActivity[];
  pagination: PaginationResults;
}

export interface FileActivitiesService {
  uri: string;
  page_size?: number;
  order_by?: string;
  order_direction?: string;
  next_page_token?: string;
  filter?: string;
}

export const FileActivityFilter = {
  all: "",
  updates: "updates",
  reads: "reads",
  my: "my",
};

export const AuditLogType = {
  server_start: 0,
  user_signup: 1,
  email_sent: 2,
  user_activated: 3,
  user_login_failed: 4,
  user_login: 5,
  user_token_refresh: 6,
  file_create: 7,
  file_rename: 8,
  set_file_permission: 9,
  entity_uploaded: 10,
  entity_downloaded: 11,
  copy_from: 12,
  copy_to: 13,
  move_to: 14,
  delete_file: 15,
  move_to_trash: 16,
  share: 17,
  share_link_viewed: 18,
  set_current_version: 19,
  delete_version: 20,
  thumb_generated: 21,
  live_photo_uploaded: 22,
  update_metadata: 23,
  edit_share: 24,
  delete_share: 25,
  mount: 26,
  relocate: 27,
  create_archive: 28,
  extract_archive: 29,
  webdav_login_failed: 30,
  webdav_account_create: 31,
  webdav_account_update: 32,
  webdav_account_delete: 33,
  payment_created: 34,
  points_change: 35,
  payment_paid: 36,
  payment_fulfilled: 37,
  payment_fulfill_failed: 38,
  storage_added: 39,
  group_changed: 40,
  user_exceed_quota_notified: 41,
  user_changed: 42,
  get_direct_link: 43,
  link_account: 44,
  unlink_account: 45,
  change_nick: 46,
  change_avatar: 47,
  membership_unsubscribe: 48,
  change_password: 49,
  enable_2fa: 50,
  disable_2fa: 51,
  add_passkey: 52,
  remove_passkey: 53,
  redeem_gift_code: 54,
  file_imported: 55,
  update_view: 56,
  delete_direct_link: 57,
  report_abuse: 58,
  oauth_grant_create: 59,
  oauth_token_exchange: 60,
  oauth_grant_revoke: 61,
};